renamed components to globalComponents because we don't want general components in that folder
