+++ src/globalComponents/LinkType.js
