+++ src/trainingarena/TrainingArena.js
@@ -1,7 +1,7 @@
 //@flow strict
 
 import * as React from 'react';
-import Navbar from '../components/Navbar.js';
+import Navbar from '../globalComponents/Navbar.js';
 
 class TrainingArena extends React.Component<{||}> {
